flask_template/templates/mongo/requirements.txt
Flask
flask-mongoengine
gunicorn

click
shortuuid
pycrypto

pandas
tablib==0.13.0
requests
